Make column matching case insensitive by default
Before: `datetime` and `dateTime` are considered different columns.

After: `datetime` and `dateTime` are considered the same column (last column prevails).

Note the output column name will still use the same casing passed in `CSVNormalizerHeader`.

`CSVNormalizer` already significantly modifies column names by normalizing to camelCase, so it is surprising the column matching is still case sensitive. It is likely that in the case a user wants `column A` and `columnA` to be the same column, that they'd also want `columna` to be the same as well.